requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.1.0
